fix delete exception and update problem

+++ src/main/java/org/elastos/hive/database/DeleteResult.java
@@ -1,7 +1,7 @@
 package org.elastos.hive.database;
 
 public class DeleteResult extends Result {
-	public long deletedCount() {
-		return (long)get("deleted_count");
+	public int deletedCount() {
+		return (int)get("deleted_count");
 	}
 }
